autonomous-shopping-assistant/requirements.txt
# Autonomous Shopping Assistant - dependencies
# Python 3.11+

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# Config & env
python-dotenv>=1.0.0

# Data validation & settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# DB
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Redis (prod cache/queue)
redis>=5.0.0

# Auth (prod JWT)
PyJWT>=2.8.0

# LLM / Agent (optional: openai for real agent)
openai>=1.12.0

# Real web search: set SERPAPI_KEY to use SerpAPI (uses httpx, no extra package)

# Observability (optional)
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
